save compatibility
Removes the audiozoo directory file path from the save so that it is valid on other systems/filepaths as long as you run the program from the home directory. There's got to be a better way to do this but I am too lazy to figure it out right now.